build: Exclude generated files from abl rule srcs
Currently, if generated files exist in the edk2 source tree, they
will be copied into the Bazel sandbox when building ABL.

In some cases, this can cause "read-only filesystem" errors when the
build attempts to overwrite the generated files. But, a more subtle
and dangerous condition can also occur.

When the user has run edksetup.sh before (outside of Bazel), a file
is generated in Conf/BuildEnv.sh that includes things like toolchain
and workspace definitions (from the host). If this file exists at
Bazel build-time, it's not overwritten - it's used. This can cause
the build to use host toolchains instead of hermetic ones which
can lead to other subtle issues.

To address this, simply exclude all generated files from the ABL
rule's "srcs" attribute. That way, no matter what state the build
environment script in the repo is, it will be generated fresh within
the Bazel sandbox.

Change-Id: I468b49f6545a23aee28085cf5b69e267225f7fa7