QcomModulePkg: Add multicore LZ4 decompression support
Add multicore decompression support for LZ4 compressed hibernation images.

If the hibernation image is compressed using the LZ4 algorithm, it will
be decompressed. If it is not compressed, no compression will be
applied. This functionality will operate regardless of the hibernation
image's compression status.
When the kernel command line parameter is set to “hibernate=nocompress,”
the SF_NOCOMPRESS_MODE flag is stored in swsusp_header->flags.
Similarly, when “hibernate.compressor=lz4” is set, the
SF_COMPRESSION_ALG_LZ4 flag is stored.

During the boot-up process, the bootloader checks the SuswspInfo->Flag
value to determine if the hibernation image is in a compressed form.

Change-Id: I2ff666e80d15914aec9c58bf08dd5fea308d3e8f