QcomModulePkg: Add support to restore image with zero pages
When generating a hibernation snapshot image, pages in memory that
consist entirely of zeros are excluded from the snapshot. To identify
such zero pages, the most significant bit (MSB) of the Page Frame Number
(PFN) is set. When the bootloader encounters a zero page, will silently
memset it to 0 and then continue on to the next page.

To provide some concrete numbers from simple ad-hoc testing, on a Hana
device we saw that out of approximately 2.36 GB of saveable pages,
0.37 GB (16%) were just zero filled and could be tracked entirely within
the packed PFN list.
As per kernel logs:
   PM: hibernation: Image created (521470 pages copied, 97590 zero pages)

Also, during our evaluation of bootloader-based hibernation, we observed
an 8% improvement in the time required for image restoration as well.

                               Time to restore
Without Zero Pages                 6.236 s
With Zero Pages                    5.634 s

Finally, such an approach could improve performance as each one of those
zero pages are not required to be loaded from the disk.

Change-Id: Idf3347044f97c1e0c4906005d4412569a219b2e1