BootLib: Add support to load images for the AutoGHGVM target
Virtualized ABL for the AutoGHGVM target uses the virtio-block
driver as the block device.
To support a trusted VM where virtio communication is through a
bounce-buffer construct, we have to initially request virtualized UEFI
to load the requested image into the bounce buffer space, and  then
copy the context from the bounce buffer space to the Guest memory.

Change-Id: I4e0b48452fad4d4dea078f5fe61ef2caa6c1a714