QcomModulePkg: Move support for package ID out of msm ID check
commit 4560c0db6079 ("QcomModulePkg: Add support for package ID
as a part of msm-ID") was checking for case when package ID was
present as part of msm ID. But in case of boards that don't care
about package ID we exit from GetPlatformMatchDtb() as package ID
is 0. So move the package ID check out of msm ID check similar to
soc rev and foundry ID check.

Change-Id: I4c471c49a5248cd0bced9252038f661595f9501c