BootLib: Add support for resetting the device state for the AutoGHGVM
The AutoGHGVM target requires ABL to run in a virtual environment
and support device lock and unlock for device reflashing purposes.
Add support for ABL to know if it needs to reset the device state
via the IsResetDeviceState flag which is set by a tool running
in the PVM context.
We also update GetBlkIOHandles() to check partition type for the
partition working over virtio-block, because ABL uses partition
type GUID to locate la_misc partition to set RECOVERY_WIPE_DATA
when resetting the device state.

Change-Id: I358652cc2915492a8af15cb390eeac1552afc7c6