edk2: QcomModulePkg: Fix compilation error when HIBERNATION_NO_AES is enabled
Resolve the compilation error that occurs when HIBERNATION_NO_AES
is enabled, and ensure that decompression functions only when the
HIBERNATION_AES flag is set. The current code assumes decompression
will only work if encryption is enabled in the kernel. This is because
the decompression logic relies on a block array, and the code for
storing this block array is written in the encryption vendor hook.
Each index of the block array contains a compressed page unit.

Change-Id: I58c56882b7e590d4b87b1f0366ccd8481bc57100